Fix RRA definition in rrd create command and parsing of rrd total result with comma
XFF was set as 0,999 when it should have been 0,999 which rrd interpreted as 0.
This worked until a new rrd version did not accept it anymore. Changed to 0.1.
Also replaced "," with "." in rrd total result. The new rrdtool uses current
numerical locale for PRINT output, which can means comma as decimal separator
which python float() does not like.