Add possibility for string comparison
suggestion to add two new commands to the calculate module 'str==' and 'str!=' which are extremely helpfull to detect for instance errors returned by the onewire plugin. If that module gets an error reading a temperature it returns the string 'error' which cannot be detected with existing commands in the calculate module.